Handle multi-scramArch in LogCollect
