+++ src/python/WMCore/WMSpec/Steps/Executors/CMSSW.py
@@ -12,13 +12,12 @@ import subprocess
 import sys
 import socket
 
-from PSetTweaks.WMTweak import readAdValues
 from WMCore.FwkJobReport.Report import addAttributesToFile
 from WMCore.WMRuntime.Tools.Scram import Scram
 from WMCore.WMSpec.Steps.Executor import Executor
 from WMCore.WMSpec.Steps.WMExecutionFailure import WMExecutionFailure
 from WMCore.WMSpec.WMStep import WMStepHelper
-from WMCore.WMRuntime.Tools.Scram import OS_TO_ARCH
+from WMCore.WMRuntime.Tools.Scram import getSingleScramArch
 
 
 def analysisFileLFN(fileName, lfnBase, job):
@@ -97,7 +96,7 @@ class CMSSW(Executor):
         logging.info('User files are %s', userFiles)
         logging.info('User sandboxes are %s', userTarball)
 
-        scramArch = self.getSingleScramArch(scramArch)
+        scramArch = getSingleScramArch(scramArch)
 
         multicoreSettings = self.step.application.multicore
         try:
@@ -330,36 +329,6 @@ class CMSSW(Executor):
             return self.errorDestination
 
         return None
-
-    @staticmethod
-    def getSingleScramArch(scramArch):
-        """
-        Figure out which scram arch is compatible with both the request and the release on the WN
-
-        Args:
-            scramArch: string or list of strings representing valid scram arches for the workflow
-
-        Returns:
-            a single scram arch
-        """
-
-        if isinstance(scramArch, list):
-            try:
-                ad = readAdValues(['glidein_required_os'], 'machine')
-                runningOS = ad['glidein_required_os'].strip('"')
-                validArches = sorted(OS_TO_ARCH[runningOS], reverse=True)
-                for requestedArch in sorted(scramArch, reverse=True):
-                    for validArch in validArches:
-                        if requestedArch.startswith(validArch):
-                            return requestedArch
-            except KeyError:
-                logging.error('OS: %s and scramArch %s do not match anything: %s', runningOS, scramArch, OS_TO_ARCH)
-                return sorted(scramArch)[-1]  # Give the most recent release if lookup fails
-
-            logging.error('scramArch %s does not match anything: %s', scramArch, OS_TO_ARCH)
-            return None
-        else:
-            return scramArch
 
 
 CONFIG_BLOB = """#!/bin/bash
